Rename user-config.ts
