+++ src/app/projects/page.tsx
@@ -4,7 +4,7 @@ import { type FC } from 'react'
 import ProjectsPage from '~/app/projects/ProjectsPage'
 import { openGraph } from '~/app/shared-metadata'
 import { type GitHubRepo, type Project } from '~/models'
-import { UserConfig } from '~/user.config'
+import { UserConfig } from '~/user-config'
 
 const PROJECTS_META_INFO: Array<{ githubRepo: string; packageUrl?: string }> = [
 	{
